+++ task2/src/components/DateRangePicker/index.jsx
@@ -1,28 +1,18 @@
-import React, { useState } from "react";
-import { addMonths, subMonths } from "date-fns";
+import React from "react";
 import CalendarHeader from "./CalendarHeader";
 import CalendarCells from "./CalendarCells";
 import "./DateRangePicker.css";
+import { useDatePicker } from "./hooks/useDatePicker";
 
 const DateRangePicker = () => {
-  const [startDate, setStartDate] = useState(null);
-  const [endDate, setEndDate] = useState(null);
-  const [currentMonth, setCurrentMonth] = useState(new Date());
-
-  const handleDateClick = (day) => {
-    if (!startDate || (startDate && endDate)) {
-      setStartDate(day);
-      setEndDate(null);
-    } else if (day >= startDate) {
-      setEndDate(day);
-    } else {
-      setStartDate(day);
-      setEndDate(startDate);
-    }
-  };
-
-  const nextMonth = () => setCurrentMonth(addMonths(currentMonth, 1));
-  const prevMonth = () => setCurrentMonth(subMonths(currentMonth, 1));
+  const {
+    currentMonth,
+    nextMonth,
+    prevMonth,
+    startDate,
+    endDate,
+    handleDateClick,
+  } = useDatePicker();
 
   return (
     <div className="calendar">
